Finish Create Expense screen UI

+++ shared/src/commonMain/kotlin/model/Finance.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Finance(
-    val month: Map<String, Double>
+    val month: Map<String, Float>
 )
